bind E in dired-mode to wdired for edits
This is a temporary with-eval-after-load hack.

Idealy, the global keybindings.el should not have mode-specific binds,
and this dired-mode bind belongs in a layer that "owns" dired.  I have
not yet made such a layer, but when I do, this patch belongs in that
layer's keybindings.el rather than here.