remove mjl/shell function
It's not working and it's not necessary to debug anyway: my bash dotfiles take
care of running within Emacs now.

For the case where this function switches to using Cygwin/bash when on Windows,
I am undecided (I'm not a heavy Windows users these days).  In the future I may
consider using the Windows Subsystem for Linux instead of Cygwin anyway (new in
Windows 10), or I might want to run Powershell (it's not such a bad shell).  Or
I would use Eshell.

Bash within Emacs on Windows is not such a big deal to me nowadays, now that I
use more civilised workstations.